sol-util: start using sol-util-impl-$(PLATFORM)
Since we were using a lot of platform #ifdefs in sol-util.c
